feat(5736): encapsulate shared chart logic into common components
