+++ app/app/analytics/private-cloud/ActiveProducts.tsx
@@ -1,5 +1,5 @@
 import { useSnapshot } from 'valtio';
-import CombinedAreaGraph from '@/components/analytics/CombinedAreaGraph';
+import LineChartCard from '@/components/analytics/LineChartCard';
 import { clusters } from '@/constants';
 import { downloadPrivateCloudActiveProducts } from '@/services/backend/analytics/private-cloud';
 import { ActiveProduct } from '@/types/analytics-private';
@@ -10,8 +10,9 @@ export default function ActiveProducts({ data }: { data: ActiveProduct[] }) {
   const pageSnapshot = useSnapshot(pageState);
   const startDate = pageSnapshot.dates?.[0] ?? data?.[0]?.date;
   const endDate = pageSnapshot.dates?.[1] ?? new Date();
+
   return (
-    <CombinedAreaGraph
+    <LineChartCard
       index="date"
       title="Active Products"
       subtitle={`This chart displays the cumulative number of products created from ${formatDate(
