+++ app/app/analytics/public-cloud/AllRequests.tsx
@@ -1,5 +1,5 @@
 import { useSnapshot } from 'valtio';
-import CombinedAreaGraph from '@/components/analytics/CombinedAreaGraph';
+import LineChartCard from '@/components/analytics/LineChartCard';
 import { downloadPublicCloudAllRequests } from '@/services/backend/analytics/public-cloud';
 import type { AllRequests } from '@/types/analytics-public';
 import { formatDate } from '@/utils/js/date';
@@ -10,7 +10,7 @@ export default function AllRequests({ data }: { data: AllRequests[] }) {
   const startDate = pageSnapshot.dates?.[0] ?? data?.[0]?.date;
   const endDate = pageSnapshot.dates?.[1] ?? new Date();
   return (
-    <CombinedAreaGraph
+    <LineChartCard
       index="date"
       title="Requests over time"
       subtitle={`This chart displays the number of requests created over time for each request type from ${formatDate(
