+++ app/app/analytics/public-cloud/ContactChangeRequests.tsx
@@ -1,5 +1,5 @@
 import { useSnapshot } from 'valtio';
-import CombinedAreaGraph from '@/components/analytics/CombinedAreaGraph';
+import LineChartCard from '@/components/analytics/LineChartCard';
 import { downloadPublicCloudContactChangeRequests } from '@/services/backend/analytics/public-cloud';
 import { ContactsChange } from '@/types/analytics-public';
 import { formatDate } from '@/utils/js/date';
@@ -10,7 +10,7 @@ export default function ContactChangeRequests({ data }: { data: ContactsChange[]
   const startDate = pageSnapshot.dates?.[0] ?? data?.[0]?.date;
   const endDate = pageSnapshot.dates?.[1] ?? new Date();
   return (
-    <CombinedAreaGraph
+    <LineChartCard
       index="date"
       title="Contact change requests over time"
       subtitle={`This chart displays edit requests where contact changes were requested and the request decision from ${formatDate(
