+++ app/app/analytics/public-cloud/RequestsDecisionTime.tsx
@@ -1,5 +1,5 @@
 import { useSnapshot } from 'valtio';
-import Histogram from '@/components/analytics/Histogram';
+import BarChartCard from '@/components/analytics/BarChartCard';
 import { downloadPublicCloudRequestsDecisionTime } from '@/services/backend/analytics/public-cloud';
 import { RequestDecisionTime } from '@/types/analytics-public';
 import { formatDate } from '@/utils/js/date';
@@ -10,7 +10,7 @@ export default function RequestsDecisionTime({ data }: { data: RequestDecisionTi
   const startDate = pageSnapshot.dates?.[0] ?? new Date('2024-01-01T00:00:00.000Z');
   const endDate = pageSnapshot.dates?.[1] ?? new Date();
   return (
-    <Histogram
+    <BarChartCard
       index="time"
       title="Request decision time frequency (%)"
       subtitle={`This chart displays the frequency of request decision times (%) for  products created from ${formatDate(
@@ -18,7 +18,6 @@ export default function RequestsDecisionTime({ data }: { data: RequestDecisionTi
       )} to ${formatDate(endDate)}.`}
       chartData={data}
       categories={['Percentage']}
-      colors={['indigo']}
       onExport={async () => downloadPublicCloudRequestsDecisionTime({ data: { ...pageSnapshot } })}
     />
   );
