+++ app/components/analytics/MultipleDoughnutChartCard.tsx
@@ -1,42 +1,13 @@
 import { Card, Title, Subtitle } from '@tremor/react';
-import type { Chart } from 'chart.js';
-import { ChartTypeRegistry, TooltipItem } from 'chart.js';
-import { ChartOptions } from 'chart.js';
 import _map from 'lodash-es/map';
 import _orderBy from 'lodash-es/orderBy';
 import _sum from 'lodash-es/sum';
 import _sumBy from 'lodash-es/sumBy';
-import { Doughnut } from 'react-chartjs-2';
-import { valueFormatter, getColor } from '@/components/analytics/helpers';
+import DoughnutChart, { DoughnutChartDataItem } from '@/components/generic/charts/DoughnutChart';
 import { formatNumber } from '@/utils/js';
 import ExportButton from '../buttons/ExportButton';
 
-const centerTextPlugin = {
-  id: 'centerText',
-  beforeDraw: (chart: Chart) => {
-    const { width, height, ctx } = chart;
-    ctx.restore();
-
-    const fontSize = (height / 150).toFixed(2);
-    ctx.font = `${fontSize}em sans-serif`;
-    ctx.textBaseline = 'middle';
-
-    const total = _sum(chart.data.datasets[0].data);
-
-    const text = total.toString();
-    const textX = Math.round((width - ctx.measureText(text).width) / 2);
-    const textY = height / 2;
-
-    ctx.fillText(text, textX, textY);
-    ctx.save();
-  },
-};
-
-interface PieGraphItem {
-  label: string;
-  value: number;
-}
-export default function PieGraph({
+export default function DoughnutChartCard({
   title,
   subtitle,
   data,
@@ -44,7 +15,7 @@ export default function PieGraph({
 }: {
   title: string;
   subtitle: string;
-  data: Record<string, PieGraphItem[]>;
+  data: Record<string, DoughnutChartDataItem[]>;
   onExport?: () => Promise<boolean>;
 }) {
   return (
@@ -54,41 +25,13 @@ export default function PieGraph({
         <Title>{title}</Title>
         <Subtitle>{subtitle}</Subtitle>
         <div className={`grid grid-cols-1 lg:grid-cols-${Object.values(data).length} lg:gap-4`}>
-          {_map(data, (items: PieGraphItem[], key: string) => {
+          {_map(data, (items: DoughnutChartDataItem[], key: string) => {
             const total = _sumBy(items, (item) => item.value);
             const orderedItems = _orderBy(items, ['value'], 'desc');
 
-            const data = {
-              labels: orderedItems.map((v) => v.label),
-              datasets: [
-                {
-                  data: orderedItems.map((v) => v.value),
-                  backgroundColor: orderedItems.map((v, ind) => getColor(ind)),
-                  hoverOffset: 10,
-                },
-              ],
-            };
-
-            const options: ChartOptions<'doughnut'> = {
-              cutout: '70%',
-              plugins: {
-                legend: {
-                  display: false,
-                },
-                tooltip: {
-                  callbacks: {
-                    label: function (context: TooltipItem<keyof ChartTypeRegistry>) {
-                      const value = valueFormatter(Number(context.raw as number));
-                      return value;
-                    },
-                  },
-                },
-              },
-            };
-
             return (
               <div className="w-full max-w-lg mx-auto" key={key}>
-                <Doughnut key={key} data={data} options={options} plugins={[centerTextPlugin]} />
+                <DoughnutChart key={key} data={orderedItems} />
                 <h4 className="text-center font-semibold mt-5">{key}</h4>
                 <ul className="tremor-List-root w-full divide-y divide-tremor-border text-tremor-content dark:divide-dark-tremor-border dark:text-dark-tremor-content">
                   {orderedItems.map((item) => {
